python/pyproject.toml
[project]
name = "literateProgrammingInConTeXt"
version = "0.0.1"
description = "Literate programming in ConTeXt - code extractor"
authors = [
    {name = "Stephen Gaito", email = "stephen@perceptisys.co.uk"},
]
dependencies = [
    "pyyaml>=6.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[project.scripts]
lpic = "lpic:cli.cli"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
